Integrate AccessControl into MetadataForm
